Cleaned up temporary files handling
Previously we created one temporary file for the inventory and one for
the ansible configuration. The home directory was set as /tmp.

The problem with this was that ansible creates a directory in /tmp and
concurrent runs of ansible will start to conflict. There will also be
an access denied error if this directory is created from ansible runs
where different users where used. Also this directory was never removed.

We now create a temporary "home" in /tmp for each ansible run. We also
put the inventory file and the ansible configuration in this directory.
After the run this whole directory gets cleaned up.