Replaced the sshagent authentication method with key
Previously if you wanted to login to the server by using a private key
you had to give the library a socket to a connected ssh-agent with the
keys loaded in it.

Now we will start the ssh-agent from inside the library and you just
enter the keys in the inventory with the key authentication method.

The agent is only started when it is needed and we will only start one
agent per ansible object instance.